+++ web/src/components/Skeleton/Animation.ts
